requirements.txt
azure-functions==1.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
idna==2.10
requests==2.24.0
selenium==3.141.0
urllib3==1.25.10
webdriver-manager==3.2.1
